Refactor Configurator component structure and imports
Moved Configurator.jsx and its CSS module into a dedicated Configurator directory. Updated all relevant imports in dependent components to reflect the new file structure. This improves project organization and maintainability.
