+++ src/components/Configurator/Configurator.jsx
@@ -2,11 +2,11 @@
 
 import { useState, Suspense } from 'react'
 import { Canvas } from '@react-three/fiber'
-import Scene from './3d/Scene'
-import LoadingFallback from './3d/LoadingFallback'
-import { CONFIG_OPTIONS } from '../config/modelConfig'
-import styles from '../styles/configurator.module.css'
-import { DEFAULT_CONFIG } from '../config/modelConfig'
+import Scene from '../3d/Scene'
+import LoadingFallback from '../3d/LoadingFallback'
+import { CONFIG_OPTIONS } from '../../config/modelConfig'
+import styles from './configurator.module.css'
+import { DEFAULT_CONFIG } from '../../config/modelConfig'
 
 export default function Configurator({ config, setConfig }) {
 	const getBackgroundForTexture = () => {
