+++ src/components/Configurator/configurator.module.css
