+++ src/components/PresetTextures.jsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { CONFIG_OPTIONS } from '../config/modelConfig'
 import { FLAVOR_CONFIG } from '../config/flavorConfig'
 import { FlavourPickerButton } from './ui/FlavourPickerButton/FlavourPickerButton'
-import styles from '../styles/configurator.module.css'
+import styles from '../components/Configurator/configurator.module.css'
 
 const PresetTextures = ({ config, setConfig }) => {
 	const handlePresetTextureChange = (textureKey) => {
